cleanup redundant scripts and update
